Added TinyPNG integration

+++ src/modules/ui/components/ImageUploader.js
@@ -10,7 +10,8 @@ export default class ImageUploader extends React.Component {
             location: 'none',
             data: null,
             bytesTransferred: 0,
-            totalBytes: 0
+            totalBytes: 0,
+            preparing: false
         }
         this.getData = this.getData.bind(this)
         this.getFileInput = () => this.fileInput || document.getElementById('fileInput')
@@ -18,6 +19,7 @@ export default class ImageUploader extends React.Component {
         this.hasValue = this.hasValue.bind(this)
         this.getExtension = this.getExtension.bind(this)
         this.random = Math.random() * 1000
+        this.getName = this.getName.bind(this)
     }
 
     getData(img) {
@@ -53,14 +55,39 @@ export default class ImageUploader extends React.Component {
         }
     }
 
+    getName(base) {
+        if (this.state.location === 'local')
+            return base + this.random + this.getExtension()
+        if (this.state.location === 'none')
+            return ''
+        return this.props.value
+    }
+
     saveImage(name) {
-        var promises = [new Promise((resolve, reject) => {
+        /*var promises = [new Promise((resolve, reject) => {
             if (this.state.location === 'local')
                 firebase.storage().ref(name + this.getExtension()).put(this.getFileInput().files[0], {cacheControl: 'no-cache'}).on(firebase.storage.TaskEvent.STATE_CHANGED, snapshot => this.setState({bytesTransferred: snapshot.bytesTransferred, totalBytes: snapshot.totalBytes}), reject, resolve)
-            else
-                resolve()
-        })]
-        if (this.props.value)
+        })]*/
+        this.setState({preparing: true})
+        var promises = []
+        if (this.state.location === 'local') {
+            promises.push(new Promise((resolve, reject) => {
+                var reader = new FileReader()
+                reader.onload = () => {
+                    var result = reader.result
+                    window.fetch('https://devgregw.com/utilities/tinify/tinify.php', {
+                        method: "POST",
+                        body: result.substr(result.lastIndexOf(',') + 1)
+                    }).then(r => {
+                        r.text().then(data => {
+                            firebase.storage().ref(name).putString(result.substr(0, result.lastIndexOf(',') + 1) + data, 'data_url', {cacheControl: 'no-cache'}).on(firebase.storage.TaskEvent.STATE_CHANGED, snapshot => this.setState({preparing: false, bytesTransferred: snapshot.bytesTransferred, totalBytes: snapshot.totalBytes}), reject, resolve)
+                        })
+                    })
+                }
+                reader.readAsDataURL(this.getFileInput().files[0])
+            }))
+        }
+        if (this.props.value && this.state.location === 'local')
             promises.push(firebase.storage().ref(this.props.value).delete())
         return Promise.all(promises)
         
@@ -95,6 +122,7 @@ export default class ImageUploader extends React.Component {
                 <Button onClick={() => this.setState({data: null, file: null})} disabled={!this.props.value} color="warning">Reset</Button>
             </ButtonGroup>
             <ProgressModal isOpen={this.state.totalBytes > 0} progressColor="primary" value={(this.state.bytesTransferred / (this.state.totalBytes || 1)) * 100} progressText={Math.round((this.state.bytesTransferred / (this.state.totalBytes || 1)) * 100) + '%'}/>
+            <ProgressModal isOpen={this.state.preparing} progressColor="primary" value={100} progressText="Preparing media..."/>
             <p>Download URL: {this.state.location.startsWith('storage') ? <a href={url} target="_blank">{url}</a> : 'none'}</p>
             <img alt="Preview" ref={i => this.getData(i || document.getElementById(`imagePreview${this.random}`))} id={`imagePreview${this.random}`} style={{width: '200px', height: 'auto', border: '1px solid black'}}/>
             </div>
